Build the initial structure of the project
There are lots of things that are being commited at once, as they
collectively outline the main structure of the project. Everything
has to be built upon this.

Migrations have not been uploaded yet, as the project is in development
phase and the queries, indexes, etc. are still being optimized.